CTF Compfest 17 Final/asdf_LAIUSDJIAS/artifact/src/web/requirements.txt
Flask==3.1.2
gunicorn==23.0.0
Jinja2==3.1.6
PyYAML==6.0.1
lxml